hammer code into controller until it works
